requirements.txt
langchain
langchain_groq
groq
# bs64
python-dotenv
streamlit